# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API de Usuários",
    "version": "1.0.0",
    "description": "API para gerenciamento de usuários"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local de desenvolvimento"
    }
  ],
  "paths": {
    "/usuarios": {
      "get": {
        "summary": "Listar todos os usuários",
        "description": "Retorna uma lista de todos os usuários cadastrados.",
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo usuário",
        "description": "Cria um novo usuário com os dados fornecidos.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}": {
      "put": {
        "summary": "Atualizar um usuário existente",
        "description": "Atualiza as informações de um usuário existente com base no ID fornecido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID do usuário a ser atualizado"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Excluir um usuário",
        "description": "Remove um usuário existente com base no ID fornecido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID do usuário a ser deletado"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário deletado com sucesso!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "6703b9b74f8d7a6d3f8ef1a1" },
          "email": { "type": "string", "example": "exemplo@dominio.com" },
          "name": { "type": "string", "example": "João Silva" },
          "age": { "type": "integer", "example": 25 }
        }
      },
      "UserInput": {
        "type": "object",
        "required": ["email", "name", "age"],
        "properties": {
          "email": { "type": "string", "example": "usuario@teste.com" },
          "name": { "type": "string", "example": "Maria Souza" },
          "age": { "type": "integer", "example": 30 }
        }
      }
    }
  }
}